Fix: 유저 관련 모델 일부 수정
